refactor: LogCardのカテゴリ定義をconstantsに移動
LogCardコンポーネント内に定義されていたカテゴリ関連の定数と関数を
`constants/learningLogs.ts`に移動し、再利用性を向上させた。

変更内容:
- `constants/learningLogs.ts` を新規作成
- LOG_CATEGORIES, getCategoryInfo, getCategoryColor を移動
- LogCard.tsx から不要なインポートを削除し、定数をインポートに変更

リファクタリングの効果:
- カテゴリ定義が一元管理される
- 他のコンポーネント（フィルターなど）でも再利用可能
- コンポーネントコードがシンプルになる（約25行削減）
